requirements.txt
openai-whisper==v20240930
